+++ nexus/lib/core/.gitignore
@@ -1,1 +1,1 @@
-libwandbcore.bin
+nexusimage.bin
